+++ posthog/queries/stickiness/stickiness.py
@@ -1,29 +1,23 @@
 import copy
 import urllib.parse
-from datetime import datetime
 from typing import Any, Dict, List
 
-from django.conf import settings
-from django.db.models.expressions import F
-from django.utils import timezone
-from sentry_sdk.api import capture_exception
-
-from ee.clickhouse.queries.stickiness.stickiness_actors import ClickhouseStickinessActors
-from ee.clickhouse.queries.stickiness.stickiness_event_query import StickinessEventsQuery
-from ee.clickhouse.sql.person import GET_LATEST_PERSON_SQL, INSERT_COHORT_ALL_PEOPLE_SQL, PERSON_STATIC_COHORT_TABLE
 from posthog.client import sync_execute
 from posthog.constants import TREND_FILTER_TYPE_ACTIONS
 from posthog.models.action import Action
-from posthog.models.cohort import Cohort
 from posthog.models.entity import Entity
 from posthog.models.filters.stickiness_filter import StickinessFilter
 from posthog.models.team import Team
 from posthog.queries.base import handle_compare
-from posthog.queries.person_distinct_id_query import get_team_distinct_ids_query
+from posthog.queries.stickiness.stickiness_actors import StickinessActors
+from posthog.queries.stickiness.stickiness_event_query import StickinessEventsQuery
 from posthog.utils import encode_get_request_params
 
 
-class ClickhouseStickiness:
+class Stickiness:
+    event_query_class = StickinessEventsQuery
+    actory_query_class = StickinessActors
+
     def run(self, filter: StickinessFilter, team: Team, *args, **kwargs) -> List[Dict[str, Any]]:
 
         response = []
@@ -36,7 +30,7 @@ class ClickhouseStickiness:
         return response
 
     def stickiness(self, entity: Entity, filter: StickinessFilter, team: Team) -> Dict[str, Any]:
-        events_query, event_params = StickinessEventsQuery(
+        events_query, event_params = self.event_query_class(
             entity, filter, team, using_person_on_events=team.actor_on_events_querying_enabled
         ).get_query()
 
@@ -52,7 +46,7 @@ class ClickhouseStickiness:
         return self.process_result(counts, filter, entity)
 
     def people(self, target_entity: Entity, filter: StickinessFilter, team: Team, request, *args, **kwargs):
-        _, serialized_actors = ClickhouseStickinessActors(entity=target_entity, filter=filter, team=team).get_actors()
+        _, serialized_actors = self.actory_query_class(entity=target_entity, filter=filter, team=team).get_actors()
         return serialized_actors
 
     def process_result(self, counts: List, filter: StickinessFilter, entity: Entity) -> Dict[str, Any]:
@@ -107,31 +101,3 @@ class ClickhouseStickiness:
                 {"filter": extra_params, "url": f"api/person/stickiness/?{urllib.parse.urlencode(parsed_params)}",}
             )
         return persons_url
-
-
-def insert_stickiness_people_into_cohort(cohort: Cohort, target_entity: Entity, filter: StickinessFilter) -> None:
-    content_sql, params = ClickhouseStickinessActors(
-        entity=target_entity, filter=filter, team=cohort.team
-    ).actor_query()
-
-    try:
-        sync_execute(
-            INSERT_COHORT_ALL_PEOPLE_SQL.format(
-                content_sql=content_sql,
-                latest_person_sql=GET_LATEST_PERSON_SQL.format(query=""),
-                cohort_table=PERSON_STATIC_COHORT_TABLE,
-                GET_TEAM_PERSON_DISTINCT_IDS=get_team_distinct_ids_query(cohort.team_id),
-            ),
-            {"cohort_id": cohort.pk, "_timestamp": datetime.now(), **params},
-        )
-        cohort.is_calculating = False
-        cohort.last_calculation = timezone.now()
-        cohort.errors_calculating = 0
-        cohort.save()
-    except Exception as err:
-        if settings.DEBUG:
-            raise err
-        cohort.is_calculating = False
-        cohort.errors_calculating = F("errors_calculating") + 1
-        cohort.save()
-        capture_exception(err)
